code improvements
I changed the structure of the combiner.py. Now we are looping over the channels outside of the run_combiner(), so we don't need to read in the input files for each channel at once. I changed the reader accordingly.

I added a progress bar for the simulations for a better overview. I also improved the reading of the sigmav values from the hdf5 files. Thermal relic density (Steigman et al.) now in the plot.