bug fixed combination of zero columns
When a zero or constant column is detected, the code is now skipping the addition/combination. Before it was just adding dummy zero columns, which caused somehow troubles. Now this is handled properly

This commit also include a minor change for the plotting. Now the plots are produced, after each channel and not at the end of the script.