enabled LklCom to run on a cluster
This commit adds a script to run LklCom on a computing farm. Therefore the run_combiner script got refactored into a script local.py and a function combiner(). local.py and cluster.py both call the combiner() function as an engine. The local.py script outputs the UL plots plus a single hdf5 file containing the limits as before and cluster.py store the results of each run in a separate hdf5 file. In addition, this commit adds a plotting script to merge the separate hdf5 files into the common hdf5 file and plot the ULs.