improved reading of J Factors
This commit add the 'Bonnivard' set of J factors in LklCom. Now the user can switch between two sets of J factors. Also the reading of the txt files and the J factors are improved, leading to an overall speed up of the code. Minor changes to the console output for a better overview.